enforcing naming convention for roxygen templates + update docs
roxygen templates naming convention i've been using is

`[field]-[name]`

where `[field]` is the roxygen field being templated (param, return), and `[name]` is the argument as called by the function (for param) or a descriptive name for other fields (like return). this way, we can more easily tell what the template stands for in the roxygen block above a function.

@wyusuf068, i hope you are OK with this convention!